(PC-17810) refactor(bookOffer): extract all function from BookingWrapper and move to new helpers folder
