+++ src/features/bookOffer/components/BookDuoChoice.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { DuoChoiceSelector } from 'features/bookOffer/components/DuoChoiceSelector'
-import { useBooking } from 'features/bookOffer/pages/BookingOfferWrapper'
+import { useBooking } from 'features/bookOffer/helpers/useBooking'
 import { Step } from 'features/bookOffer/pages/reducer'
 import { TouchableOpacity } from 'ui/components/TouchableOpacity'
 import { Spacer, Typo } from 'ui/theme'
