+++ src/features/bookOffer/components/BookingDetails.native.test.tsx
@@ -5,7 +5,8 @@ import { mocked } from 'ts-jest/utils'
 import { navigate, useRoute } from '__mocks__/@react-navigation/native'
 import { OfferStockResponse } from 'api/gen'
 import { mockDigitalOffer, mockOffer } from 'features/bookOffer/fixtures/offer'
-import { useBookingOffer, useBookingStock } from 'features/bookOffer/pages/BookingOfferWrapper'
+import { useBookingOffer } from 'features/bookOffer/helpers/useBookingOffer'
+import { useBookingStock } from 'features/bookOffer/helpers/useBookingStock'
 import { BookingState, initialBookingState } from 'features/bookOffer/pages/reducer'
 import { offerStockResponseSnap } from 'features/offer/fixtures/offerStockResponse'
 import { useIsUserUnderage } from 'features/profile/helpers/useIsUserUnderage'
