+++ src/features/bookOffer/components/BookingEventChoices.tsx
@@ -7,7 +7,7 @@ import { BookDateChoice } from 'features/bookOffer/components/BookDateChoice'
 import { BookDuoChoice } from 'features/bookOffer/components/BookDuoChoice'
 import { BookHourChoice } from 'features/bookOffer/components/BookHourChoice'
 import { BookingDetails } from 'features/bookOffer/components/BookingDetails'
-import { useBooking } from 'features/bookOffer/pages/BookingOfferWrapper'
+import { useBooking } from 'features/bookOffer/helpers/useBooking'
 import { Step } from 'features/bookOffer/pages/reducer'
 import { useCreditForOffer } from 'features/offer/helpers/useHasEnoughCredit/useHasEnoughCredit'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
