+++ src/features/bookOffer/components/BookingImpossible.tsx
@@ -2,8 +2,8 @@ import { useNavigation } from '@react-navigation/native'
 import React, { useEffect } from 'react'
 import styled from 'styled-components/native'
 
-import { useBooking } from 'features/bookOffer/pages/BookingOfferWrapper'
-import { useNotifyWebappLinkSent } from 'features/bookOffer/services/useNotifyWebappLinkSent'
+import { useNotifyWebappLinkSent } from 'features/bookOffer/api/useNotifyWebappLinkSent'
+import { useBooking } from 'features/bookOffer/helpers/useBooking'
 import { useAddFavorite, useFavorite } from 'features/favorites/api'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/firebase/analytics'
