+++ src/features/bookOffer/helpers/useBooking.tsx
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+
+import { BookingContext } from '../pages/BookingContext'
+import { IBookingContext } from '../pages/BookingOfferWrapper'
+
+export const useBooking = (): IBookingContext => {
+  return useContext(BookingContext)
+}
