+++ src/features/bookOffer/pages/BookingOfferModal.tsx
@@ -1,13 +1,12 @@
 import React, { useEffect } from 'react'
 
+import { useBooking } from 'features/bookOffer/helpers/useBooking'
+import { useModalContent } from 'features/bookOffer/helpers/useModalContent'
+import { BookingWrapper } from 'features/bookOffer/pages/BookingWrapper'
 import { analytics } from 'libs/firebase/analytics'
 import { AppModal } from 'ui/components/modals/AppModal'
 import { ModalLeftIconProps } from 'ui/components/modals/types'
 import { Close } from 'ui/svg/icons/Close'
-
-import { useModalContent } from '../services/useModalContent'
-
-import { BookingWrapper, useBooking } from './BookingOfferWrapper'
 
 interface Props {
   visible: boolean
