+++ src/features/bookOffer/pages/__mocks__/useBooking.tsx
@@ -1,0 +1,11 @@
+export const useBooking = jest.fn().mockReturnValue({
+  bookingState: {
+    offerId: undefined,
+    stockId: undefined,
+    step: 1,
+    quantity: 2,
+    date: undefined,
+  },
+  dispatch: () => null,
+  dismissModal: () => null,
+})
