+++ src/features/bookOffer/pages/__mocks__/useBookingOffer.tsx
@@ -1,0 +1,6 @@
+export const useBookingOffer = jest.fn().mockReturnValue({
+  isDigital: false,
+  name: 'mon nom',
+  stocks: [],
+  venue: {},
+})
