+++ src/libs/hooks/__tests__/useShowReview.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import InAppReview from 'react-native-in-app-review'
 import waitForExpect from 'wait-for-expect'
 
-import { useReviewInAppInformation } from 'features/bookOffer/services/useReviewInAppInformation'
+import { useReviewInAppInformation } from 'features/bookOffer/helpers/useReviewInAppInformation'
 import * as useFeatureFlag from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { useShowReview } from 'libs/hooks/useShowReview'
 import { render } from 'tests/utils'
