+++ src/libs/hooks/useShowReview.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import InAppReview from 'react-native-in-app-review'
 
-import { useReviewInAppInformation } from 'features/bookOffer/services/useReviewInAppInformation'
+import { useReviewInAppInformation } from 'features/bookOffer/helpers/useReviewInAppInformation'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { eventMonitoring } from 'libs/monitoring'
